getting rid of shorts removes more compiler warnings

+++ src/c-game/rendertext.c
@@ -5,7 +5,7 @@
 
 
 
-short char_coords[96][4] =
+int char_coords[96][4] =
 {
         {0,0,25,64},            //!
         {25,0,54,64},      //"
